+++ src/utility/system_info.ts
@@ -19,7 +19,7 @@ export const SystemInfo = {
     name: "AntOS",
   },
   application: {
-    version: "7.1.21",
+    version: "7.1.22",
     shortVersionString: ShortVersion.v7,
     name: "Declarative AI",
   },
